requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.6
django-discover-runner==0.4
django-debug-toolbar==1.0.1
Sphinx==1.2b1
termcolor==1.1.0
sqlparse
